@@ -1,10 +1,8 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: [
-    "./src/**/*.{js,jsx,ts,tsx}",
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
+    content: ['./src/**/*.{js,jsx,ts,tsx}'],
+    theme: {
+        extend: {}
+    },
+    plugins: []
 }
